+++ lib/an_lifecycle_cancellable.dart
@@ -6,6 +6,7 @@ export 'dialog/dialog_ext.dart';
 export 'dialog/navigator_ext.dart';
 export 'lifecycle/lifecycle_async.dart';
 export 'lifecycle/lifecycle_context_effect.dart';
+export 'lifecycle/lifecycle_controller.dart';
 export 'lifecycle/lifecycle_data.dart';
 export 'lifecycle/lifecycle_effect.dart';
 export 'lifecycle/lifecycle_ext.dart';