+++ lib/lifecycle/lifecycle_context_effect.dart
@@ -22,28 +22,24 @@ class _LLauncherObserver with LifecycleStateChangeObserver {
   LifecycleState _lastState = LifecycleState.destroyed;
 
   _LLauncherObserver._(BuildContext context)
-      : _context = WeakReference(context) {}
+      : _context = WeakReference(context);
 
   @override
   void onStateChange(LifecycleOwner owner, LifecycleState state) {
-    if (state == LifecycleState.destroyed && launchOnDestroy != null) {
-      launchOnDestroy!(owner.lifecycle);
+    if (state == LifecycleState.destroyed) {
+      launchOnDestroy?.call(owner.lifecycle);
+      _lastState = state;
+      return;
     }
 
     if (_context.target == null || _context.target?.mounted != true) return;
-    if (_firstCreate &&
-        state == LifecycleState.created &&
-        launchOnFirstCreate != null) {
+    if (_firstCreate && state == LifecycleState.created) {
       _firstCreate = false;
-      launchOnFirstCreate!(owner.lifecycle);
-    } else if (_firstStart &&
-        state == LifecycleState.started &&
-        launchOnFirstStart != null) {
+      launchOnFirstCreate?.call(owner.lifecycle);
+    } else if (_firstStart && state == LifecycleState.started) {
       _firstStart = false;
-      launchOnFirstStart!(owner.lifecycle);
-    } else if (_firstResume &&
-        state == LifecycleState.resumed &&
-        launchOnFirstResume != null) {
+      launchOnFirstStart?.call(owner.lifecycle);
+    } else if (_firstResume && state == LifecycleState.resumed) {
       _dRunOnResume(owner);
     }
     if (_lastState < state) {
@@ -69,7 +65,7 @@ class _LLauncherObserver with LifecycleStateChangeObserver {
       /// 特殊情况下会resume触发在build之前故将此事件推迟
       if (l.currentState > LifecycleState.destroyed &&
           _context.target?.mounted == true) {
-        launchOnFirstResume!(owner.lifecycle);
+        launchOnFirstResume?.call(owner.lifecycle);
       }
     });
   }
@@ -89,28 +85,24 @@ class _DLauncherObserver<T extends Object> with LifecycleStateChangeObserver {
   LifecycleState _lastState = LifecycleState.destroyed;
 
   _DLauncherObserver._(BuildContext context, this._data)
-      : _context = WeakReference(context) {}
+      : _context = WeakReference(context);
 
   @override
   void onStateChange(LifecycleOwner owner, LifecycleState state) {
-    if (state == LifecycleState.destroyed && launchOnDestroy != null) {
-      launchOnDestroy!(owner.lifecycle, _data);
+    if (state == LifecycleState.destroyed) {
+      launchOnDestroy?.call(owner.lifecycle, _data);
+      _lastState = state;
+      return;
     }
 
     if (_context.target == null || _context.target?.mounted != true) return;
-    if (_firstCreate &&
-        state == LifecycleState.created &&
-        launchOnFirstCreate != null) {
+    if (_firstCreate && state == LifecycleState.created) {
       _firstCreate = false;
-      launchOnFirstCreate!(owner.lifecycle, _data);
-    } else if (_firstStart &&
-        state == LifecycleState.started &&
-        launchOnFirstStart != null) {
+      launchOnFirstCreate?.call(owner.lifecycle, _data);
+    } else if (_firstStart && state == LifecycleState.started) {
       _firstStart = false;
-      launchOnFirstStart!(owner.lifecycle, _data);
-    } else if (_firstResume &&
-        state == LifecycleState.resumed &&
-        launchOnFirstResume != null) {
+      launchOnFirstStart?.call(owner.lifecycle, _data);
+    } else if (_firstResume && state == LifecycleState.resumed) {
       _dRunOnResume(owner);
     }
     if (_lastState < state) {
@@ -136,7 +128,7 @@ class _DLauncherObserver<T extends Object> with LifecycleStateChangeObserver {
       /// 特殊情况下会resume触发在build之前故将此事件推迟
       if (l.currentState > LifecycleState.destroyed &&
           _context.target?.mounted == true) {
-        launchOnFirstResume!(owner.lifecycle, _data);
+        launchOnFirstResume?.call(owner.lifecycle, _data);
       }
     });
   }
@@ -144,6 +136,8 @@ class _DLauncherObserver<T extends Object> with LifecycleStateChangeObserver {
 
 final _withLifecycleKey = Object();
 final _withLifecycleDataKey = Object();
+final _withLifecycleAndDataKey = Object();
+final _withLifecycleAndExtDataKey = Object();
 
 class _BuildContextLifecycleWithDataKey {
   final Object? key;
@@ -210,6 +204,7 @@ extension BuildContextLifecycleWithExt on BuildContext {
 
   /// 从当前的Context中获取Lifecycle使用 并且data 同属于 key的一部分
   /// 如果使用factory 则必须保证多次调用时返回同一值 否则将会视为新建
+  @Deprecated('use withLifecycleAndDataEffect')
   T withLifecycleEffectData<T extends Object>({
     T? data,
     T Function()? factory,
@@ -276,5 +271,169 @@ extension BuildContextLifecycleWithExt on BuildContext {
       observer.launchOnDestroy = launchOnDestroy;
     }
     return data;
+  }
+
+  /// 从当前的Context中获取Lifecycle使用 并且data 同属于 key的一部分
+  /// 如果使用factory 则必须保证多次调用时返回同一值 否则将会视为新建
+  T withLifecycleAndDataEffect<T extends Object>({
+    T? data,
+    T Function()? factory,
+    T Function(Lifecycle lifecycle)? factory2,
+    DLauncher<T>? launchOnFirstCreate,
+    DLauncher<T>? launchOnFirstStart,
+    DLauncher<T>? launchOnFirstResume,
+    DLauncher<T>? repeatOnStarted,
+    DLauncher<T>? repeatOnResumed,
+    DLauncher<T>? launchOnDestroy,
+    Object? key,
+  }) {
+    assert(data != null || factory != null || factory2 != null,
+        'data and factory and factory2 cannot be null at the same time');
+
+    data ??= factory?.call();
+
+    final lifecycle = Lifecycle.of(this);
+
+    data ??= factory2?.call(lifecycle);
+
+    if (data == null) {
+      throw 'data and factory and factory2 cannot be null at the same time';
+    }
+
+    final ctx = this;
+    final d = data;
+
+    final cache = lifecycle.extData
+        .putIfAbsent<Map<BuildContext, Map<Object, _DLauncherObserver>>>(
+            key: _withLifecycleAndDataKey, ifAbsent: () => WeakHashMap());
+
+    final cache2 = cache.putIfAbsent(ctx, () {
+      final result = HashMap<Object, _DLauncherObserver>();
+      lifecycle.addLifecycleObserver(_WeakMapClear(result));
+      return result;
+    });
+
+    final observer =
+        cache2.putIfAbsent(key == null ? d : _LifecycleAndDataKey(d, key), () {
+      final observer = _DLauncherObserver<T>._(ctx, d);
+      observer.launchOnFirstCreate = launchOnFirstCreate;
+      observer.launchOnFirstStart = launchOnFirstStart;
+      observer.launchOnFirstResume = launchOnFirstResume;
+      observer.launchOnDestroy = launchOnDestroy;
+      if (repeatOnStarted != null) {
+        observer._repeatOn[LifecycleState.started] = repeatOnStarted;
+      }
+      if (repeatOnResumed != null) {
+        observer._repeatOn[LifecycleState.resumed] = repeatOnResumed;
+      }
+
+      lifecycle.addObserver(observer);
+      return observer;
+    }) as _DLauncherObserver<T>;
+
+    if (repeatOnStarted != null) {
+      observer._repeatOn[LifecycleState.started] = repeatOnStarted;
+    }
+    if (repeatOnResumed != null) {
+      observer._repeatOn[LifecycleState.resumed] = repeatOnResumed;
+    }
+    if (launchOnDestroy != null) {
+      observer.launchOnDestroy = launchOnDestroy;
+    }
+    return data;
+  }
+
+  /// 从当前context生成一个绑定到liveDate内部的缓存数据
+  /// context 销毁时自动清理
+  /// 与 [withLifecycleAndExtDataEffect] 数据相同
+  T withLifecycleExtData<T extends Object>({
+    T Function()? factory,
+    T Function(Lifecycle lifecycle)? factory2,
+    Object? key,
+  }) {
+    final lifecycle = Lifecycle.of(this);
+    factory ??= () => factory2!(lifecycle);
+
+    final Map<BuildContext, Map<Object, Object?>> contextData =
+        lifecycle.extData.putIfAbsent(
+            key: _withLifecycleAndExtDataKey, ifAbsent: WeakHashMap.new);
+
+    final data = contextData.putIfAbsent(this, () {
+      final result = HashMap<Object, Object?>();
+
+      /// 不持有 Map，防止内存泄漏
+      lifecycle.addLifecycleObserver(_WeakMapClear(result));
+      return result;
+    });
+
+    return data.putIfAbsent(key == null ? T : TypedKey<T>(key), factory) as T;
+  }
+
+  /// 从当前context生成一个绑定到liveDate内部的缓存数据 并提供生命周期的相关函数
+  /// context 销毁时自动清理
+  /// 与 [withLifecycleExtData] 数据相同
+  T withLifecycleAndExtDataEffect<T extends Object>({
+    T Function()? factory,
+    T Function(Lifecycle lifecycle)? factory2,
+    DLauncher<T>? launchOnFirstCreate,
+    DLauncher<T>? launchOnFirstStart,
+    DLauncher<T>? launchOnFirstResume,
+    DLauncher<T>? repeatOnStarted,
+    DLauncher<T>? repeatOnResumed,
+    DLauncher<T>? launchOnDestroy,
+    Object? key,
+  }) {
+    final data = withLifecycleExtData(
+      factory: factory,
+      factory2: factory2,
+      key: key,
+    );
+    return withLifecycleAndDataEffect(
+      data: data,
+      launchOnFirstCreate: launchOnFirstCreate,
+      launchOnFirstStart: launchOnFirstStart,
+      launchOnFirstResume: launchOnFirstResume,
+      repeatOnStarted: repeatOnStarted,
+      repeatOnResumed: repeatOnResumed,
+      launchOnDestroy: launchOnDestroy,
+    );
+  }
+}
+
+class _WeakMapClear with LifecycleStateChangeObserver {
+  final WeakReference<Map> _weakReference;
+
+  _WeakMapClear(Map map) : _weakReference = WeakReference(map);
+
+  void clear() {
+    _weakReference.target?.clear();
+  }
+
+  @override
+  void onStateChange(LifecycleOwner owner, LifecycleState state) {
+    if (state > LifecycleState.destroyed && _weakReference.target == null) {
+      // 如果发现已经被销毁了 则将当前的Observer也注销掉
+      owner.removeLifecycleObserver(this, fullCycle: false);
+    } else if (state == LifecycleState.destroyed) {
+      clear();
+    }
+  }
+}
+
+class _LifecycleAndDataKey {
+  final Object data;
+  final Object? key;
+
+  _LifecycleAndDataKey(this.data, this.key);
+
+  @override
+  int get hashCode => Object.hash(data, key);
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is _LifecycleAndDataKey &&
+        other.data == data &&
+        other.key == key;
   }
 }